refactor: moved  spam_check #5
Issues when the bot tries to load the spam_check
as a cog which stops the detect.py cog from
 propperly loading
